// src/types/statistical/statistical.ts
export type StatisticalPrice = {
  code_shop: string;
  date_today: string;
  date_yesterday: string;
  date_yesterday_1: string;
  date_lastMonth: string;
  date_lastMonth2: string;
  date_lastMonth_1: string;
  date_lastMonth_2: string;
};
export type StatisticalFUll = {
  code_shop: string;
  year: string;
};